Work around windows path length limitations by moving immutable -> full in open source (new generator paths) and shortening long file names.
Re-enable windows bazel test.

PiperOrigin-RevId: 630443328
