+++ src/google/protobuf/compiler/java/file.cc
@@ -25,7 +25,7 @@
 #include "google/protobuf/compiler/java/generator_common.h"
 #include "google/protobuf/compiler/java/generator_factory.h"
 #include "google/protobuf/compiler/java/helpers.h"
-#include "google/protobuf/compiler/java/immutable/generator_factory.h"
+#include "google/protobuf/compiler/java/full/generator_factory.h"
 #include "google/protobuf/compiler/java/lite/generator_factory.h"
 #include "google/protobuf/compiler/java/name_resolver.h"
 #include "google/protobuf/compiler/java/options.h"
