+++ src/google/protobuf/compiler/java/full/enum.h
