+++ src/google/protobuf/compiler/java/full/generator_factory.h
