+++ src/google/protobuf/compiler/java/full/map_field.cc
@@ -5,7 +5,7 @@
 // license that can be found in the LICENSE file or at
 // https://developers.google.com/open-source/licenses/bsd
 
-#include "google/protobuf/compiler/java/immutable/map_field.h"
+#include "google/protobuf/compiler/java/full/map_field.h"
 
 #include <string>
