+++ src/google/protobuf/compiler/java/full/message.h
@@ -17,7 +17,7 @@
 #include <vector>
 
 #include "google/protobuf/compiler/java/generator_factory.h"
-#include "google/protobuf/compiler/java/immutable/field_generator.h"
+#include "google/protobuf/compiler/java/full/field_generator.h"
 #include "google/protobuf/descriptor.h"
 
 namespace google {
