+++ src/google/protobuf/compiler/java/full/primitive_field.h
@@ -14,7 +14,7 @@
 
 #include <string>
 
-#include "google/protobuf/compiler/java/immutable/field_generator.h"
+#include "google/protobuf/compiler/java/full/field_generator.h"
 #include "google/protobuf/descriptor.h"
 
 namespace google {
