+++ src/google/protobuf/compiler/java/lite/make_field_gens.h
@@ -5,12 +5,15 @@
 // license that can be found in the LICENSE file or at
 // https://developers.google.com/open-source/licenses/bsd
 
-#ifndef GOOGLE_PROTOBUF_COMPILER_JAVA_IMMUTABLE_MAKE_FIELD_GENERATORS_H__
-#define GOOGLE_PROTOBUF_COMPILER_JAVA_IMMUTABLE_MAKE_FIELD_GENERATORS_H__
+#ifndef GOOGLE_PROTOBUF_COMPILER_JAVA_LITE_MAKE_FIELD_GENS_H__
+#define GOOGLE_PROTOBUF_COMPILER_JAVA_LITE_MAKE_FIELD_GENS_H__
+
+#include <memory>
+#include <vector>
 
 #include "google/protobuf/compiler/java/context.h"
 #include "google/protobuf/compiler/java/generator_common.h"
-#include "google/protobuf/compiler/java/immutable/field_generator.h"
+#include "google/protobuf/compiler/java/lite/field_generator.h"
 #include "google/protobuf/descriptor.h"
 
 namespace google {
@@ -18,7 +21,7 @@ namespace protobuf {
 namespace compiler {
 namespace java {
 
-FieldGeneratorMap<ImmutableFieldGenerator> MakeImmutableFieldGenerators(
+FieldGeneratorMap<ImmutableFieldLiteGenerator> MakeImmutableFieldLiteGenerators(
     const Descriptor* descriptor, Context* context);
 
 }  // namespace java
@@ -26,4 +29,4 @@ FieldGeneratorMap<ImmutableFieldGenerator> MakeImmutableFieldGenerators(
 }  // namespace protobuf
 }  // namespace google
 
-#endif  // GOOGLE_PROTOBUF_COMPILER_JAVA_IMMUTABLE_MAKE_FIELD_GENERATORS_H__
+#endif  // GOOGLE_PROTOBUF_COMPILER_JAVA_LITE_MAKE_FIELD_GENS_H__
