+++ src/google/protobuf/compiler/java/lite/message_builder.cc
@@ -20,7 +20,7 @@
 #include "google/protobuf/compiler/java/field_common.h"
 #include "google/protobuf/compiler/java/helpers.h"
 #include "google/protobuf/compiler/java/lite/enum.h"
-#include "google/protobuf/compiler/java/lite/make_field_generators.h"
+#include "google/protobuf/compiler/java/lite/make_field_gens.h"
 #include "google/protobuf/compiler/java/name_resolver.h"
 #include "google/protobuf/descriptor.pb.h"
 #include "google/protobuf/io/printer.h"
