+++ src/google/protobuf/compiler/objectivec/text_format_decode_data_unittest.cc
@@ -5,12 +5,12 @@
 // license that can be found in the LICENSE file or at
 // https://developers.google.com/open-source/licenses/bsd
 
-#include "google/protobuf/compiler/objectivec/text_format_decode_data.h"
 
 #include <cstdint>
 #include <string>
 
 #include <gtest/gtest.h>
+#include "google/protobuf/compiler/objectivec/tf_decode_data.h"
 
 // Must be included last
 #include "google/protobuf/port_def.inc"
